Added list and term resources and related classes

+++ src/main/java/ca/canada/digital/search/assessment/dao/LanguageDao.java
@@ -16,7 +16,7 @@ public class LanguageDao extends AbstractDAO<Language> {
         super(factory);
     }
 
-    public Optional<Language> findById(Long id) {
+    public Optional<Language> findById(Integer id) {
         return Optional.ofNullable(get(id));
     }
